+++ CHANGELOG.md
@@ -14,3 +14,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - referenceDataStore class
 - AutoSaveControlledForm component
 - Login and Logout button components
+- Client HTTP service